# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_dgrco_quikslate_internal_domain.LRole:
    enum:
    - employee
    - manager
    - location_lead
    - ""
    - admin
    type: string
    x-enum-comments:
      EmptyRole: no location-scoped role (identity-only sessions)
      LocationLead: manages (leads) all roles at a location
      Manager: manages employees
    x-enum-descriptions:
    - ""
    - manages employees
    - manages (leads) all roles at a location
    - no location-scoped role (identity-only sessions)
    - ""
    x-enum-varnames:
    - Employee
    - Manager
    - LocationLead
    - EmptyRole
    - Admin
...
  github_com_dgrco_quikslate_internal_response.errorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_handler.BusinessDTO:
    properties:
      business_id:
        type: string
      created_at:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  internal_handler.BusinessMemberDetailDTO:
    properties:
      business_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      is_admin:
        type: boolean
      is_primary_admin:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  internal_handler.CreateBusinessResponse:
    properties:
      business_id:
        type: string
    type: object
...
  internal_handler.MultipleBusinessMemberDetailResponse:
    properties:
      members:
        items:
          $ref: '#/definitions/internal_handler.BusinessMemberDetailDTO'
        type: array
    type: object
  internal_handler.MultipleBusinessResponse:
    properties:
      businesses:
        items:
          $ref: '#/definitions/internal_handler.BusinessDTO'
        type: array
    type: object
...
  internal_handler.MyBusinessMembershipResponse:
    properties:
      is_admin:
        type: boolean
      is_primary_admin:
        type: boolean
      location_roles:
        items:
          $ref: '#/definitions/internal_handler.MyLocationRoleDTO'
        type: array
      user_id:
        type: string
    type: object
  internal_handler.MyLocationRoleDTO:
    properties:
      location_id:
        type: string
      role:
        $ref: '#/definitions/github_com_dgrco_quikslate_internal_domain.LRole'
    type: object
...
  internal_handler.SimpleResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_handler.createBusinessRequest:
    properties:
      business_name:
        type: string
    type: object
...
  internal_handler.renameBusinessRequest:
    properties:
      business_name:
        type: string
    type: object
...
  internal_handler.setAdminRequest:
    properties:
      admin:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Scheduling API for businesses, locations, positions, shifts, and employees.
  title: QuikSlate API
  version: "1.0"
paths:
...
  /businesses:
    post:
      consumes:
      - application/json
      description: Create a new business. The caller becomes its primary admin.
      parameters:
      - description: Business details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createBusinessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.CreateBusinessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Create business
      tags:
      - businesses
  /businesses/{businessId}:
    delete:
      description: Delete a business and all of its data (locations, positions, shifts,
        members). Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete business
      tags:
      - businesses
    get:
      description: Get a business by ID. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.BusinessDTO'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Get business
      tags:
      - businesses
    patch:
      consumes:
      - application/json
      description: Rename a business. Admin only.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: New name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.renameBusinessRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Rename business
      tags:
      - businesses
...
  /businesses/{businessId}/members:
    get:
      description: Get business member details by business ID
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultipleBusinessMemberDetailResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Get business member details
      tags:
      - businesses
  /businesses/{businessId}/members/{userId}:
    delete:
      description: Remove a user from the business (and all of their location roles/position
        assignments). Admin only; admins cannot remove other admins.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Remove business member
      tags:
      - businesses
  /businesses/{businessId}/members/{userId}/admin:
    patch:
      consumes:
      - application/json
      description: Grant or revoke admin status for a business member. Admin to grant
        admin status to a non-admin member; only the primary admin can modify another
        admin's status.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      - description: Target user ID
        in: path
        name: userId
        required: true
        type: string
      - description: Desired admin status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.setAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.SimpleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Set business member admin status
      tags:
      - businesses
  /businesses/{businessId}/members/me:
    get:
      description: Get the caller's own admin status and location roles within this
        business. Lets the frontend decide what admin-only or location-scoped UI to
        show, without probing each action individually.
      parameters:
      - description: Business ID
        in: path
        name: businessId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MyBusinessMembershipResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Get my business membership
      tags:
      - businesses
...
  /businesses/me:
    get:
      description: Retrieves all businesses that the currently authenticated user
        belongs to.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.MultipleBusinessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_dgrco_quikslate_internal_response.errorResponse'
      security:
      - BearerAuth: []
      summary: Get Businesses By User
      tags:
      - businesses
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"